refactor: version is now an internal package
Like for previous commits, just refactoring/yak shaving.
