+++ Readme.md
@@ -105,7 +105,7 @@ To update bundled binary data use:
 
 1. update binary data as described above;
 
-2. update `version/version.go`;
+2. update `internal/version/version.go`;
 
 3. make sure you have updated dependencies;
