+++ internal/cli/root/root.go
@@ -7,7 +7,7 @@ import (
 	"github.com/ooni/probe-cli/internal/log/handlers/batch"
 	"github.com/ooni/probe-cli/internal/log/handlers/cli"
 	"github.com/ooni/probe-cli/internal/utils"
-	"github.com/ooni/probe-cli/version"
+	"github.com/ooni/probe-cli/internal/version"
 )
 
 // Cmd is the root command
